requirements.txt
# Web Framework and HTTP
Flask==2.0.1
requests==2.31.0
Werkzeug==2.0.1

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Financial Data
yfinance==0.2.31
websocket-client==1.6.1

# Environment Management
python-dotenv==1.0.0

# Database
SQLAlchemy==1.4.41
psycopg2-binary==2.9.5

# Additional Dependencies
plotly==5.13.0
dash==2.8.1